chore: Remove references to official domain

+++ src/backend/accounts/progress.ts
@@ -2,6 +2,7 @@ import { ofetch } from "ofetch";
 
 import { getAuthHeaders } from "@/backend/accounts/auth";
 import { ProgressResponse } from "@/backend/accounts/user";
+import { BACKEND_URL } from "@/setup/constants";
 import { AccountWithToken } from "@/stores/auth";
 import { ProgressMediaItem, ProgressUpdateItem } from "@/stores/progress";
 
@@ -103,6 +104,7 @@ export async function removeProgress(
   episodeId?: string,
   seasonId?: string,
 ) {
+  if (!BACKEND_URL) return;
   await ofetch(`/users/${account.userId}/progress/${id}`, {
     method: "DELETE",
     headers: getAuthHeaders(account.token),
